requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
certifi==2020.6.20
chardet==3.0.4
click==8.0.1
conda==4.3.16
distlib==0.3.1
filelock==3.0.12
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pipenv==2020.8.13
pycosat==0.6.3
pytz==2021.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.16.0
SQLAlchemy==1.4.22
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
Werkzeug==2.0.1

uuid~=1.30
PyJWT~=2.1.0